third_party/blink/web_tests/flag-specific/highdpi/external/wpt/html/rendering/non-replaced-elements/the-hr-element-0/setting-overflow-visible-expected.txt

This is a testharness.js-based test.
[FAIL] control
  assert_equals: divAfter.offsetTop expected 52 but got 51
[FAIL] overflow: visible
  assert_equals: divAfter.offsetTop expected 2 but got 1
Harness: the test ran to completion.
